feat(user, auth): Get logged in user informations
* feat(auth): add UnauthorizedNoTokenError and authentication check in AuthService

This commit introduces a new error class, UnauthorizedNoTokenError, to handle cases where no authorization token is provided. Additionally, it implements the throwErrorIfNotAuthenticated method in AuthService to validate the presence of a token and verify its validity.

* refactor(auth): rename throwErrorIfNotAuthenticated to guardUserIsAuthenticated and return user

This commit renames the method throwErrorIfNotAuthenticated to guardUserIsAuthenticated for clarity and updates its functionality to return the authenticated user instead of throwing an error directly. The UnauthorizedInvalidTokenError import has also been removed as it is no longer used.

* feat(auth): add /user endpoint and update token parsing in auth service

This commit updates the AuthService to correctly retrieve the authorization token from the request headers. It also modifies the UserController to utilize the AuthService for user authentication and adds a new UsersController for handling user-related routes. The UserModule is updated to include the new controller and service dependencies.

* refactor(auth): sanitize user response and update controller return types

This commit introduces a new method in AuthService to sanitize user responses by omitting the password field. The UserController is updated to reflect this change, ensuring that the getUser method returns a user object without the password. Additionally, the commented-out cookie handling code in AuthController has been removed for clarity.

* refactor(factory): improve route path construction for clarity

This commit refactors the route path construction in the Application class to enhance readability. The logic for building the full path has been simplified, ensuring that the leading slash is correctly applied only when necessary, while maintaining the existing functionality.